pyproject.toml
[tool.poetry]
name = "thndr-be-task"
version = "1.0.0"
description = ""
authors = ["Ali <ali@thndr.app>"]
readme = "README.md"
packages = [
    { include = "app" }
]

[tool.poetry.dependencies]
python = "^3.12"
fastapi = {extras = ["standard"], version = "^0.112.2"}
SQLAlchemy = "^2.0.32"
alembic = "^1.13.2"
option = "2.1.0"
slowapi = "0.1.9"
httpx = "^0.27.2"

[tool.poetry.dev-dependencies]
ipdb = "^0.13.13"

[tool.poetry.group.dev.dependencies]
ruff = "^0.7.1"
pre-commit = "^4.0.1"


[tool.poetry.group.test.dependencies]
pytest-asyncio = "^0.24.0"
aiosqlite = "^0.20.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
